Move ParkingBannerForm to shared for reuse
